+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [3.4.0] - 2021-06-15
+- Updated to work in a SSL environment.
+- Updated to use the mysql_libs v5.2.0 library.
+
+### Added
+- add_ssl:  Add SSL options to the dump command line.
+- Added -l option to allow for the use of SSL connections.
+
+### Changed
+- run_program:  Call add_ssl to process SSL options if detected and added check not to run database dump command if SSL configuration is not setup correctly.
+- config/mysql_cfg.py.TEMPLATE:  Added SSL options.
+
+
 ## [3.3.2] - 2021-05-04
 - Updated to work in MySQL 8.0 environment.
 - Updated to use the mysql_libs v5.1.0 library.